Made collision pixel perfect.
Fixed bugs in genetics (calling wrong random function).

Made it so that you can dynamically configure the amount of asteroids that a network will consider - the network calculates it's input size and hidden neuron count statically based on the  ASTEROIDS_TO_CONSIDER constant in Simulation.kt.

Added vector input for the network. Rather than using quadrants which didn't give enough to learn from. Network input is now N vectors to the nearest N asteroids. Currently the vector is not calculated with the "wrap around" property of the asteroids game. So an asteriod may be on the top of the screen and the ship at the bottom - the true distance between them is small but the algorithm currently thinks it on the other side of the screen.

Worked a lot on the fitness function - A networks fitness is now a function of the ratio of asteroids destroyed multiplied by the factor of time it took to do so. I.e. otpimizing for most asteriods destroyed in least amount of time.